requirements.txt
# Faster-Whisper ASR 模块依赖

# 核心依赖
numpy>=1.24.0
librosa>=0.10.0
soundfile>=0.12.0
pyyaml>=6.0.0
loguru>=0.7.0

# ASR 核心库
faster-whisper>=0.10.0

# 可选：GPU 加速支持
# torch>=2.0.0
# torchaudio>=2.0.0

# 开发/测试（可选）
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# pytest-cov>=4.1.0
